make an abstract middleware
